+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/calendar/bottomsheet/EventsPerDayBottomSheetFragment.kt
@@ -5,14 +5,13 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.databinding.DataBindingUtil
-import androidx.fragment.app.viewModels
+import androidx.hilt.navigation.fragment.hiltNavGraphViewModels
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.fragment.findNavController
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.teameetmeet.meetmeet.R
 import com.teameetmeet.meetmeet.databinding.FragmentEventsPerDayBottomSheetBinding
-import com.teameetmeet.meetmeet.presentation.calendar.CalendarFragmentDirections
-import com.teameetmeet.meetmeet.presentation.calendar.CalendarViewModel
+import com.teameetmeet.meetmeet.presentation.calendar.monthcalendar.MonthCalendarViewModel
 import com.teameetmeet.meetmeet.presentation.model.EventSimple
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.flow.collectLatest
@@ -23,7 +22,7 @@ class EventsPerDayBottomSheetFragment : BottomSheetDialogFragment(), EventItemCl
     private var _binding: FragmentEventsPerDayBottomSheetBinding? = null
     private val binding get() = requireNotNull(_binding)
 
-    private val viewModel: CalendarViewModel by viewModels({ requireParentFragment() })
+    private val viewModel: MonthCalendarViewModel by hiltNavGraphViewModels(R.id.nav_graph_calendar)
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
@@ -54,8 +53,8 @@ class EventsPerDayBottomSheetFragment : BottomSheetDialogFragment(), EventItemCl
             eventsPerDayBsBtnAddEvent.setOnClickListener {
                 viewModel.currentDate.value.date?.let { date ->
                     requireParentFragment().findNavController().navigate(
-                        CalendarFragmentDirections
-                            .actionCalendarFragmentToAddEventActivity(date)
+                        EventsPerDayBottomSheetFragmentDirections
+                            .actionEventsPerDayBottomSheetFragmentToAddEventActivity(date)
                     )
                 }
             }
@@ -69,7 +68,8 @@ class EventsPerDayBottomSheetFragment : BottomSheetDialogFragment(), EventItemCl
 
     override fun onItemClick(eventSimple: EventSimple) {
         requireParentFragment().findNavController().navigate(
-            CalendarFragmentDirections.actionCalendarFragmentToEventStoryActivity(eventSimple.id)
+            EventsPerDayBottomSheetFragmentDirections
+                .actionEventsPerDayBottomSheetFragmentToEventStoryActivity(eventSimple.id)
         )
     }
 }
