+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/calendar/monthcalendar/CalendarEvent.kt
@@ -1,4 +1,4 @@
-package com.teameetmeet.meetmeet.presentation.calendar
+package com.teameetmeet.meetmeet.presentation.calendar.monthcalendar
 
 import com.teameetmeet.meetmeet.presentation.model.EventSimple
 import java.io.Serializable
