+++ android/app/src/main/java/com/teameetmeet/meetmeet/presentation/calendar/monthcalendar/CalendarItemClickListener.kt
@@ -1,4 +1,4 @@
-package com.teameetmeet.meetmeet.presentation.calendar
+package com.teameetmeet.meetmeet.presentation.calendar.monthcalendar
 
 import com.teameetmeet.meetmeet.presentation.model.CalendarItem
