+++ android/app/src/main/res/navigation/nav_graph_home.xml
@@ -11,9 +11,6 @@
         android:label="calendar_fragment"
         tools:layout="@layout/fragment_calendar">
         <action
-            android:id="@+id/action_calendar_fragment_to_add_event_activity"
-            app:destination="@id/add_event_activity" />
-        <action
             android:id="@+id/action_calendar_fragment_to_setting_activity"
             app:destination="@id/setting_activity" />
         <action
@@ -22,9 +19,6 @@
         <action
             android:id="@+id/action_calendar_fragment_to_notification_activity"
             app:destination="@id/notification_activity" />
-        <action
-            android:id="@+id/action_calendar_fragment_to_eventStoryActivity"
-            app:destination="@id/event_story_activity" />
         <action
             android:id="@+id/action_calendar_fragment_to_follow_fragment"
             app:destination="@id/follow_fragment" />
@@ -52,16 +46,6 @@
         tools:layout="@layout/fragment_group" />
 
     <activity
-        android:id="@+id/add_event_activity"
-        android:name="com.teameetmeet.meetmeet.presentation.addevent.AddEventActivity"
-        android:label="add_event_activity"
-        tools:layout="@layout/activity_add_event">
-        <argument
-            android:name="date"
-            app:argType="java.time.LocalDate" />
-    </activity>
-
-    <activity
         android:id="@+id/setting_activity"
         android:name="com.teameetmeet.meetmeet.presentation.setting.SettingActivity"
         android:label="setting_activity"
@@ -79,13 +63,4 @@
         android:label="Notification_activity"
         tools:layout="@layout/activity_notification" />
 
-    <activity
-        android:id="@+id/event_story_activity"
-        android:name="com.teameetmeet.meetmeet.presentation.eventstory.EventStoryActivity"
-        android:label="event_story_activity"
-        tools:layout="@layout/activity_event_story" >
-        <argument
-            android:name="eventId"
-            app:argType="integer" />
-    </activity>
 </navigation>
